feat(package)!: the repo is now a unity package (#106)
* move every files in correct folder

* added & rename asmdef

* fix readme

* Update package.json

* fix: assembly definition

* feat!: change the folders architecture

* Create .gitignore

* feat: update the readme to explain how to install

* fix: remove deeplink WIP
