--- Scenario/package.json
@@ -1,10 +1,0 @@
-{
-    "name": "com.scenario.editor",
-    "displayName": "Scenario Unity Plugin",
-    "version": "0.3.0",
-    "unity": "2022.3",
-    "dependencies": {
-        "com.unity.editorcoroutines": "1.0.0",
-        "com.unity.nuget.newtonsoft-json": "2.0.0"
-    }
-}
