+++ package/Documentation/Scenario Knowledge Center - Unity Workflows.url
@@ -1,0 +1,6 @@
+[InternetShortcut]
+URL=https://help.scenario.com/workflow-tutorials#unity-plugin
+IDList=
+HotKey=0
+IconFile=C:\Users\fangh\AppData\Local\Mozilla\Firefox\Profiles\3v5d162j.default-release\shortcutCache\szF4o2i34bUZPRW73S4a3w==.ico
+IconIndex=0
