+++ package/Editor/Images/ImageDataClasses.cs
