+++ package/Editor/Plugins/PluginSettings.cs
