+++ package/Editor/Plugins/RestSharp/RestSharpException.cs
