+++ package/package.json
@@ -1,0 +1,21 @@
+{
+    "name": "com.scenarioinc.scenario",
+    "displayName": "Scenario",
+    "description": "Craft unique and style-consistent game assets with custom-trained AI models.",
+    "documentationUrl": "https://help.scenario.com/",
+    "changelogUrl": "https://github.com/scenario-labs/Scenario-Unity/",
+    "licensesUrl": "https://github.com/scenario-labs/Scenario-Unity/",
+    "version": "0.3.0",
+    "unity": "2022.3",
+    "dependencies": {
+        "com.unity.editorcoroutines": "1.0.0",
+        "com.unity.nuget.newtonsoft-json": "2.0.0"
+    },
+    "samples": [
+        {
+            "displayName": "Unity Editor Layout",
+            "description": "Example layout for 2D workflow",
+            "path": "Samples/UnityEditorLayout"
+        }
+    ]
+}
